+++ src/commandGroups/othershells.ts
@@ -1,11 +1,13 @@
 import { commandTable as evalZ} from "../commands/eval.js";
 import { commandTable as cmd } from "../commands/cmd.js";
+import { commandTable as powershell } from "../commands/powershell.js";
 import { cmdTableToCommandCompounts } from "../tools/commandCreatorTools.js";
 import { commandCollection } from "../tools/commandCollection.js";
 
 const commandTable = {
     ...evalZ,
-    ...cmd
+    ...cmd,
+    ...powershell
 };
 
 const compounds = cmdTableToCommandCompounts(commandTable);