+++ changelog.md
@@ -15,6 +15,8 @@ changes:
   * getonlinedataservice -> allows fetching data from the web
   * json -> allows converting from and to json strings
   * internal -> internal command that is not included by default
+  * powershell -> allows you to execute powershell commands
+  * timer -> allows you to manage session timers
 * commands were reintroduced:
   * info -> it has been mistakely removed from the preview versions
 * commands were rewritten: